+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### Toyorm v0.6.0-alpha (Dec 28 2018)
 - [NEW] add MustAddr method for Toy
 - [CHANGE] In Insert operation, CreatedAt field will keep the value unchanged when it is not zero
 - [FIX] invoke IsZero with not struct value will panic